Use an internal dataset
This required moving the data import into a single file, so I also deleted the raw csv since it seemed like it was only needed for communicating between the two files.

Since I was in here, I also tidied the code up a little to avoid the readr messages and the dplyr warning.